refactor: Move content_generation module into core
