+++ crates/sherwood/src/core/content_generation.rs
