+++ crates/sherwood/src/core/mod.rs
@@ -1,2 +1,3 @@
+pub mod content_generation;
 pub mod markdown_config;
 pub mod utils;
